requirements.txt
torch
diffusers
transformers
Pillow
numpy
matplotlib
huggingface_hub
